refactored cytoscape into smaller modules
