+++ web/src/web/pages/sitrep/callbacks/cytoscape.py
@@ -1,244 +1,14 @@
 import dash
-import pandas as pd
-import requests
 import warnings
 from dash import Input, Output, State, callback, callback_context
-from datetime import datetime
 from typing import Tuple, Any
 
-from models.census import CensusRow
-from web.config import get_settings
 from web.pages.sitrep import CAMPUSES, ids
 from web import SITREP_DEPT2WARD_MAPPING
-from web.stores import ids as store_ids
-from web.utils import Timer
 
 from web.logger import logger
 
 logger.info("Preparing cytoscape map")
-
-
-@callback(
-    Output(ids.DEPTS_OPEN_STORE, "data"),
-    Input(ids.DEPT_GROUPER, "value"),
-    Input(store_ids.DEPT_STORE, "data"),
-)
-def _store_depts(dept_grouper: str, depts: list[dict]) -> list[dict]:
-    """Need a list of departments for ALL_ICUS or this campus"""
-    if dept_grouper == "ALL_ICUS":
-        icus = SITREP_DEPT2WARD_MAPPING.keys()
-        these_depts = [dept for dept in depts if dept.get("department") in icus]
-    else:
-        try:
-            these_depts = [
-                dept for dept in depts if dept.get("location_name") == dept_grouper
-            ]
-        except TypeError as e:
-            print(e)
-            warnings.warn(f"No departments found at {dept_grouper} campus")
-            these_depts = []
-    return these_depts
-
-
-@callback(
-    Output(ids.DEPTS_OPEN_STORE_NAMES, "data"),
-    Input(ids.DEPTS_OPEN_STORE, "data"),
-)
-def _dept_open_store_names(depts_open: list[dict]) -> list[str]:
-    return [i.get("department", {}) for i in depts_open]
-
-
-@callback(
-    Output(ids.ROOMS_OPEN_STORE, "data"),
-    Input(ids.DEPTS_OPEN_STORE, "data"),
-    Input(store_ids.ROOM_STORE, "data"),
-)
-def _store_rooms(
-    depts: list[dict],
-    rooms: list[dict],
-) -> list[dict]:
-    """Need a list of rooms for this building"""
-    dfdepts = pd.DataFrame.from_records(depts)
-    dfdepts = dfdepts[["department", "hl7_department"]]
-    dfrooms = pd.DataFrame.from_records(rooms)
-    # default inner join drops rooms not in the selected departments
-    dfrooms = dfrooms.merge(dfdepts, on="department")
-    # drop closed rooms
-    dfrooms = dfrooms.loc[~dfrooms["closed"], :]
-
-    return dfrooms.to_dict(orient="records")  # type: ignore
-
-
-@callback(
-    Output(ids.BEDS_STORE, "data"),
-    Input(ids.DEPTS_OPEN_STORE, "data"),
-    Input(ids.ROOMS_OPEN_STORE, "data"),
-    Input(store_ids.BEDS_STORE, "data"),
-)
-def _store_beds(
-    depts: list[dict],
-    rooms: list[dict],
-    beds: list[dict],
-) -> list[dict]:
-    """
-    Return a list of beds using the local filtered versions of depts/rooms
-    - generate the floor_index from the bed_number to permit appropriate sorting
-    """
-
-    bedsdf = pd.DataFrame.from_records(beds)
-    dfdepts = pd.DataFrame.from_records(depts)
-    dfrooms = pd.DataFrame.from_records(rooms)
-
-    dfdepts = dfdepts[["department", "floor_order"]]
-
-    # drop beds where rooms are closed
-    # look for bays where all beds are closed
-    dft = bedsdf.groupby("hl7_room")["closed"].all()
-    dft = pd.DataFrame(dft).reset_index()
-    dft.rename(columns={"closed": "closed_all_beds"}, inplace=True)
-    dfrooms = dfrooms.merge(dft, on="hl7_room")
-
-    # now close a room if any of the following are true
-    dfrooms["closed"] = dfrooms["closed"] | dfrooms["closed_all_beds"]
-    dfrooms.drop(columns=["closed_all_beds"], inplace=True)
-    # drop closed rooms
-    dfrooms = dfrooms.loc[~dfrooms["closed"], :]
-    dfrooms = dfrooms[["hl7_room", "is_sideroom"]]
-
-    # inner join to drop rooms without beds
-    bedsdf = bedsdf.merge(dfrooms, on="hl7_room", how="inner")
-    # inner join to drop closed_perm_01
-    bedsdf = bedsdf.merge(dfdepts, on="department", how="inner")
-
-    bedsdf = bedsdf[bedsdf["bed_number"] != -1]
-    # bedsdf = bedsdf[~bedsdf["closed"]]
-
-    def _gen_floor_indices(df: pd.DataFrame) -> pd.DataFrame:
-        # now generate floor_y_index
-        df.sort_values(
-            ["floor", "floor_order", "department", "bed_number"], inplace=True
-        )
-        floor_depts = df[["floor", "floor_order", "department"]].drop_duplicates()
-        floor_depts.sort_values(["floor", "floor_order"], inplace=True)
-        floor_depts["floor_y_index"] = floor_depts.reset_index().index + 1
-        df = df.merge(floor_depts, how="left")
-
-        # create a floor x_index by sorting and ranking within floor_y_index
-        df.sort_values(["floor_y_index", "bed_number"], inplace=True)
-        df["floor_x_index"] = df.groupby("floor_y_index")["bed_number"].rank(
-            method="first", na_option="keep"
-        )
-        df.sort_values(["location_string"], inplace=True)
-        return df
-
-    bedsdf = _gen_floor_indices(bedsdf)
-
-    res: list[dict] = bedsdf.to_dict(orient="records")
-    return res
-
-
-@callback(
-    Output(ids.CENSUS_STORE, "data"),
-    Input(ids.DEPT_GROUPER, "value"),
-    Input(ids.DEPTS_OPEN_STORE_NAMES, "data"),
-)
-def _store_census(
-    dept_grouper: str,
-    depts_open_names: list[str],
-) -> list[dict]:
-    """
-    Store CensusRow as list of dictionaries after filtering out closed
-    departments for that building
-    Args:
-        dept_grouper: one of UCH/WMS/GWB/NHNN
-        depts_open_names: list of departments that are open
-
-    Returns:
-        Filtered list of CensusRow dictionaries
-
-    """
-    if dept_grouper == "ALL_ICUS":
-        response = requests.get(
-            f"{get_settings().api_url}/census/",
-            params={"departments": SITREP_DEPT2WARD_MAPPING.keys()},
-        )
-    else:
-        campus_short_name = next(
-            i.get("label") for i in CAMPUSES if i.get("value") == dept_grouper
-        )
-
-        response = requests.get(
-            f"{get_settings().api_url}/census/campus/",
-            params={"campuses": campus_short_name},
-        )
-
-    res = [CensusRow.parse_obj(row).dict() for row in response.json()]
-    res = [row for row in res if row.get("department") in depts_open_names]
-    return res
-
-
-@callback(
-    Output(ids.SITREP_STORE, "data"),
-    Input(ids.DEPT_SELECTOR, "value"),
-    Input(store_ids.SITREP_STORE, "data"),
-    prevent_initial_callback=True,
-    background=True,
-)
-def _store_sitrep(dept: str, sitreps: dict) -> Any:
-    """
-    Args:
-        dept: the department
-        sitreps: dictionary of sitreps
-
-    Returns:
-        additonal patient level data
-
-    """
-    # ward2dept = {v:k for k,v in SITREP_DEPT2WARD_MAPPING.items()}
-    # department = ward2dept.get(dept)
-    ward = SITREP_DEPT2WARD_MAPPING.get(dept)
-    if not ward:
-        warnings.warn(f"No sitrep data available for {ward}")
-        return [{}]
-    return sitreps[ward]
-
-
-@Timer(text="Discharge updates: Elapsed time: {:.3f} seconds")
-def _get_discharge_updates(delta_hours: int = 48) -> list[dict]:
-    response = requests.get(
-        f"{get_settings().api_url}/baserow/discharge_status",
-        params={"delta_hours": delta_hours},
-    )
-    if response.status_code == 200:
-        return response.json()  # type: ignore
-    else:
-        warnings.warn("No data found for discharge statues (from Baserow " "store)")
-        return [{}]
-
-
-@callback(
-    Output(ids.DISCHARGES_STORE, "data"),
-    Input(ids.DEPT_SELECTOR, "value"),
-)
-def store_discharge_status(dept: str) -> list[dict]:
-    """
-    Get discharge status
-    Deduplicate to most recent only
-    Refreshes on ward update
-    """
-    if not dept:
-        return dash.no_update  # type: ignore
-    discharges = _get_discharge_updates(delta_hours=36)
-    if not discharges or not len(discharges):
-        return dash.no_update  # type: ignore
-    df = pd.DataFrame.from_records(discharges)
-    try:
-        df.sort_values(["csn", "modified_at"], ascending=[True, False], inplace=True)
-        df.drop_duplicates(["csn"], inplace=True)
-    except KeyError as e:
-        print(e)
-        warnings.warn("Unable to sort: is the dataframe empty?")
-    return df.to_dict(orient="records")  # type: ignore
 
 
 @callback(
@@ -519,42 +289,3 @@ def _prepare_cyto_elements_ward(
         return dash.no_update  # type: ignore
 
     return elements  # type: ignore
-
-
-def format_census(census: dict) -> dict:
-    """Given a census object return a suitably formatted dictionary"""
-    mrn = census.get("mrn", "")
-    encounter = str(census.get("encounter", ""))
-    lastname = census.get("lastname", "").upper()
-    firstname = census.get("firstname", "").title()
-    initials = (
-        f"{census.get('firstname', '?')[0]}"
-        f""
-        f""
-        f""
-        f"{census.get('lastname', '?')[0]}"
-    )
-    date_of_birth = census.get("date_of_birth", "1900-01-01")  # default dob
-    dob = datetime.fromisoformat(date_of_birth)
-    dob_fshort = datetime.strftime(dob, "%d-%m-%Y")
-    dob_flong = datetime.strftime(dob, "%d %b %Y")
-    age = int((datetime.utcnow() - dob).days / 365.25)
-    sex = census.get("sex")
-    if sex is None:
-        sex = ""
-    else:
-        sex = "M" if sex.lower() == "m" else "F"
-
-    return dict(
-        mrn=mrn,
-        encounter=encounter,
-        lastname=lastname,
-        firstname=firstname,
-        initials=initials,
-        dob=dob,
-        dob_fshort=dob_fshort,
-        dob_flong=dob_flong,
-        age=age,
-        sex=sex,
-        demographic_slug=f"{firstname} {lastname} | {age}{sex} | MRN {mrn}",
-    )
