+++ web/src/web/pages/sitrep/icus.py
@@ -5,7 +5,7 @@ import json
 from dash import dcc, html
 from pathlib import Path
 
-import web.pages.sitrep.callbacks.bedlist  # noqa
+import web.pages.sitrep.layouts.bedlist  # noqa
 
 # noqa suppresses black errors when linting since you need this import for
 # access to callbacks
@@ -13,9 +13,12 @@ import web.pages.sitrep.callbacks.cytoscape  # noqa
 import web.pages.sitrep.callbacks.discharges  # noqa
 import web.pages.sitrep.callbacks.inspector  # noqa
 import web.pages.sitrep.callbacks.widgets  # noqa
+import web.pages.sitrep.callbacks.sitrep  # noqa
+import web.pages.sitrep.callbacks.census  # noqa
+import web.pages.sitrep.callbacks.beds  # noqa
 from web.pages.sitrep import ids
 from web import SITREP_DEPT2WARD_MAPPING
-from web.style import colors, replace_colors_in_stylesheet
+from web.style import replace_colors_in_stylesheet
 
 dash.register_page(__name__, path="/sitrep/icus", name="Critical Care")
